% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_classifications.R
\name{combine_classifications}
\alias{combine_classifications}
\title{Combine hierarchical classifications}
\usage{
combine_classifications(..., sep = ".")
}
\arguments{
\item{...}{\verb{[list]} A collection of lists, one for each classification,
each giving
the "digits" that represent each level in the hierarchy, as made
by \code{\link[=expand_classification]{expand_classification()}} or \code{\link[=split_classification]{split_classification()}}.}

\item{sep}{\verb{[character(1)]} A character used to separate the classifications
in \code{...}. The
default separates levels across classifications by \code{"."}.}
}
\value{
A list with a entry for each level in the combined classification.
}
\description{
Combine hierarchical classifications by stacking one classification after
another.
}
\examples{
# Combine an unbalanced industry classification with a balanced
# geographic classification.
industry <- c("111", "112", "12")
region <- c("11", "21", "22")

combine_classifications(
  expand_classification(industry, pad = "0"),
  expand_classification(region)
)
}
