% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_structure-class.R
\name{is_aggregation_structure}
\alias{is_aggregation_structure}
\title{Test if an object is an aggregation structure}
\usage{
is_aggregation_structure(x)
}
\arguments{
\item{x}{\verb{[object]} An object to test.}
}
\value{
Returns \code{TRUE} if \code{x} inherits from \code{\link{piar_aggregation_structure}}.
}
\description{
Test if an object is a price index aggregation structure.
}
