% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.piar_index.R
\name{levels.piar_index}
\alias{levels.piar_index}
\alias{levels<-.piar_index}
\alias{set_levels}
\title{Get the levels for a price index}
\usage{
\method{levels}{piar_index}(x)

\method{levels}{piar_index}(x) <- value

set_levels(x, value)
}
\arguments{
\item{x}{\verb{[piar_index]} A price index, as made by,
e.g., \code{\link[=elementary_index]{elementary_index()}}.}

\item{value}{\verb{[character]} A character vector, or something that can be
coerced into one, giving the replacement levels for \code{x}.}
}
\value{
\code{levels()} returns a character vector with the levels for a price index.

The replacement method returns a copy of \code{x} with the levels in \code{value}.
(\code{set_levels()} is an alias that's easier to use with pipes.)
}
\description{
Methods to get and set the levels for a price index.
}
\seealso{
Other index methods:
\code{\link[=[.piar_index]{[.piar_index()}},
\code{\link{aggregate.piar_index}},
\code{\link[=as.data.frame.piar_index]{as.data.frame.piar_index()}},
\code{\link[=as.ts.piar_index]{as.ts.piar_index()}},
\code{\link[=chain]{chain()}},
\code{\link[=contrib]{contrib()}},
\code{\link[=head.piar_index]{head.piar_index()}},
\code{\link[=is.na.piar_index]{is.na.piar_index()}},
\code{\link{mean.piar_index}},
\code{\link[=merge.piar_index]{merge.piar_index()}},
\code{\link[=split.piar_index]{split.piar_index()}},
\code{\link[=stack.piar_index]{stack.piar_index()}},
\code{\link[=time.piar_index]{time.piar_index()}},
\code{\link[=window.piar_index]{window.piar_index()}}
}
\concept{index methods}
