% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piar-package.R
\docType{data}
\name{price_data}
\alias{price_data}
\alias{ms_prices}
\alias{ms_weights}
\alias{fs_prices}
\alias{fs_weights}
\title{Price data}
\description{
Sample price and weight data for both a match sample and fixed sample type
index.
}
