% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute_weights.R
\name{transmute_weights}
\alias{transmute_weights}
\title{Transmute weights for a generalized mean}
\usage{
transmute_weights(x, weights = NULL, order = 0, to = 1, mean = NA)
}
\arguments{
\item{x}{A strictly positive numeric vector.}

\item{weights}{A strictly positive numeric vector of weights, the same length
as \code{x}. The default is to equally weight each element of \code{x}.}

\item{order}{A finite number giving the order of the generalized mean.}

\item{to}{A finite number giving the order of the target generalized mean.}

\item{mean}{A finite number giving the generalized mean of \code{x} and \code{weights},
if known. The default computes this values.}
}
\value{
A numeric vector, the same length as \code{x}, that sums to 1.
}
\description{
Transmute weight to turn a generalized mean of a given order into a
generalized mean of order \code{to}. Useful for calculating additive and
multiplicative decompositions for generalized-mean indexes.
}
\examples{
x <- 1:3
w <- 3:1

# Calculate the geometric mean as an arithmetic mean.
gmean(x, order = 0)
gmean(x, transmute_weights(x, order = 0, to = 1), order = 1)
}
\references{
See \code{vignette("decomposing-indexes")} for more details.
}
\seealso{
Other math functions:
\code{\link[=gmean]{gmean()}},
\code{\link[=nested_gmean]{nested_gmean()}},
\code{\link[=scale_weights]{scale_weights()}},
\code{\link[=transmute_weights2]{transmute_weights2()}},
\code{\link[=update_weights]{update_weights()}}
}
\concept{math functions}
