% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmean.R
\name{gmean}
\alias{gmean}
\title{Generalized mean}
\usage{
gmean(x, weights = NULL, r = 1, na.rm = FALSE)
}
\arguments{
\item{x}{A strictly positive numeric vector.}

\item{weights}{A strictly positive numeric vector of weights, the same length
as \code{x}. The default is to equally weight each element of \code{x}.}

\item{r}{A finite number giving the order of the generalized mean.}
}
\value{
A number.
}
\description{
Calculated a weighted generalized mean
}
