% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_structure-class.R
\name{is_aggregation_structure}
\alias{is_aggregation_structure}
\title{Test if an object is a price index aggregation structure}
\usage{
is_aggregation_structure(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Returns \code{TRUE} if \code{x} inherits from \code{\link{piar_aggregation_structure}}
}
\description{
Test if an object is a price index aggregation structure
}
\seealso{
Other aggregation structure methods: 
\code{\link{update.piar_aggregation_structure}()},
\code{\link{weights.piar_aggregation_structure}()}
}
\concept{aggregation structure methods}
