% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmean.R
\name{nested_gmean}
\alias{nested_gmean}
\title{Nested generalized means}
\usage{
nested_gmean(
  x,
  weights = list(NULL, NULL),
  order = c(1, -1),
  outer_weights = NULL,
  outer_order = 0,
  na.rm = FALSE
)
}
\arguments{
\item{x}{A strictly positive numeric vector.}

\item{weights}{A list of strictly positive numeric vector of weights, each
the same length as \code{x}, for each of the inner generalized means. \code{NULL}
elements of \code{weights} equally weight each element of \code{x}. The default
uses equal weights for each inner generalized mean.}

\item{order}{A finite numeric vector giving the order of each of the inner
generalized means. The default computes an arithmetic mean and a harmonic
mean.}

\item{outer_weights}{A strictly positive numeric vector weights for each of
the inner generalized means as used in the outer generalized mean. The
default weights each inner generalized mean equally.}

\item{outer_order}{A finite number giving the order of the outer generalized
mean.}

\item{na.rm}{Should missing values in \code{x} and \code{weights} be removed? By
default missing values in \code{x} or \code{weights} return a missing value.}
}
\value{
A numeric value for the nested generalized mean.
}
\description{
Calculate a weighted (outer) generalized mean of two (inner) generalized
means.
}
\examples{
# example code

}
\seealso{
Other math functions:
\code{\link[=gmean]{gmean()}},
\code{\link[=scale_weights]{scale_weights()}},
\code{\link[=transmute_weights]{transmute_weights()}},
\code{\link[=transmute_weights2]{transmute_weights2()}},
\code{\link[=update_weights]{update_weights()}}
}
\concept{math functions}
