% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piar-package.R
\docType{package}
\name{piar-package}
\alias{piar}
\alias{piar-package}
\title{piar: Price Index Aggregation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Most price indexes are made with a two-step procedure, where period-over-period elemental indexes are first calculated for a collection of elemental aggregates at each point in time, and then aggregated according to a price index aggregation structure. These indexes can then be chained together to form a time series that gives the evolution of prices with respect to a fixed base period. This package contains a collection of functions that revolve around this work flow, making it easy to build standard price indexes, and implement the methods described by Balk (2008, \doi{10.1017/CBO9780511720758}), von der Lippe (2007, \doi{10.3726/978-3-653-01120-3}), and the CPI manual (2020, \doi{10.5089/9781484354841.069}) for bilateral price indexes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marberts.github.io/piar/}
  \item \url{https://github.com/marberts/piar}
  \item Report bugs at \url{https://github.com/marberts/piar/issues}
}

}
\author{
\strong{Maintainer}: Steve Martin \email{marberts@protonmail.com} (\href{https://orcid.org/0000-0003-2544-9480}{ORCID}) [copyright holder]

}
\keyword{internal}
