% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-class.R
\name{piar_index}
\alias{piar_index}
\alias{chainable_piar_index}
\alias{direct_piar_index}
\title{Price index objects}
\description{
There are several classes to represent price indexes.
\itemize{
\item All indexes inherit from the \code{piar_index} virtual class.
\item Period-over-period indexes that can be chained over time inherit from
\code{chainable_piar_index}.
\item Fixed-base indexes inherit from \code{direct_piar_index}.
}
}
\details{
The \code{piar_index} object is a list-S3 class with the following
components:
\describe{
\item{index}{A matrix of index values with a column for each period in \code{time}
and a row for each level in \code{levels}.}
\item{contrib}{A list-matrix containing named vectors that give the
percent-change contributions for each price relative with a column for each
time period in \code{time} and a row for each level in \code{levels}, or \code{NULL}.}
\item{levels}{A character vector giving the levels of the index.}
\item{time}{A character vector giving the time periods for the index.}
}

The \code{chainable_piar_index} and \code{direct_piar_index} subclasses have
the same structure as the \code{piar_index} class, but differ in the methods
used to manipulate the indexes.
}
