\name{price data}
\alias{ms_prices}
\alias{ms_weights}
\alias{fs_prices}
\alias{fs_weights}
\docType{data}

\title{
Price data
}

\description{
Sample price and weight data for both a 'match sample' and 'fixed sample' type index.
}

\usage{
data("ms_prices")

data("ms_weights")

data("fs_prices")

data("fs_weights")
}
