% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-class.R
\name{time.piar_index}
\alias{time.piar_index}
\alias{time<-}
\alias{time<-.piar_index}
\alias{start.piar_index}
\alias{end.piar_index}
\title{Get the time periods for a price index}
\usage{
\method{time}{piar_index}(x, ...)

time(x) <- value

\method{time}{piar_index}(x) <- value

\method{start}{piar_index}(x, ...)

\method{end}{piar_index}(x, ...)
}
\arguments{
\item{x}{A price index, as made by, e.g., \code{\link[=elemental_index]{elemental_index()}}.}

\item{...}{Further arguments passed to or used by methods.}

\item{value}{A character vector, or something that can be coerced into one,
giving the replacement time periods for \code{x}.}
}
\value{
\code{time()} return a character vector with the time periods for a price index.
\code{start()} and \code{end()} return the first and last time period.

The replacement method returns a copy of \code{x} with the time periods in
\code{value}.
}
\description{
Methods to get and set the time periods for a price index.
}
\seealso{
Other index methods: 
\code{\link{[.piar_index}()},
\code{\link{aggregate.piar_index}()},
\code{\link{as.data.frame.piar_index}()},
\code{\link{chain}()},
\code{\link{contrib}()},
\code{\link{head.piar_index}()},
\code{\link{is.na.piar_index}()},
\code{\link{levels.piar_index}()},
\code{\link{mean.piar_index}()},
\code{\link{merge.piar_index}()},
\code{\link{stack.piar_index}()},
\code{\link{vcov.aggregate_piar_index}()}
}
\concept{index methods}
