% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute_weights.R
\name{transmute_weights2}
\alias{transmute_weights2}
\title{Transmute weights for a nested generalized mean}
\usage{
transmute_weights2(
  x,
  weights = list(NULL, NULL),
  order = c(1, -1),
  outer_weights = c(1, 1),
  outer_order = 0,
  to = 1,
  pivot = outer_order
)
}
\arguments{
\item{x}{A strictly positive numeric vector.}

\item{weights}{A list of strictly positive numeric vector of weights, each
the same length as \code{x}, for each of the inner generalized means. \code{NULL}
elements of \code{weights} equally weight each element of \code{x}. The default
uses equal weights for each inner generalized mean.}

\item{order}{A finite numeric vector giving the order of each of the inner
generalized means. The default computes an arithmetic mean and a harmonic
mean.}

\item{outer_weights}{A strictly positive numeric vector weights for each of
the inner generalized means as used in the outer generalized mean. The
default weights each inner generalized mean equally.}

\item{outer_order}{A finite number giving the order of the outer generalized
mean.}

\item{to}{A finite number giving the order of the target generalized mean for
the transmuted weights. The default constructs weights for an arithmetic
mean.}

\item{pivot}{A finite number giving the pivot value for the transmuted
weights. The default uses the order of the outer generalized mean,
otherwise \code{to} is common alternative.
See \code{vignette("decomposing-indexes")} for details.}
}
\value{
A numeric vector, the same length as \code{x}, that sums to 1.
}
\description{
Transmute weights to turn a nested generalized mean of a given order into a
generalized mean of order \code{to}. Useful for calculating additive and
multiplicative decompositions for an index made
of nested generalized means (e.g., Fisher index).
}
\examples{
x <- 1:3
w1 <- 3:1
w2 <- c(1, 2, 1)

# Calculate the geometric mean of the arithmetic and harmonic means
# as an arithmetic mean.
nested_gmean(x, list(w1, w2))
gmean(x, transmute_weights2(x, list(w1, w2), to = 1))
}
\references{
See \code{vignette("decomposing-indexes")} for more details.
}
\seealso{
Other math functions:
\code{\link[=gmean]{gmean()}},
\code{\link[=nested_gmean]{nested_gmean()}},
\code{\link[=scale_weights]{scale_weights()}},
\code{\link[=transmute_weights]{transmute_weights()}},
\code{\link[=update_weights]{update_weights()}}
}
\concept{math functions}
