% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_weights.R
\name{update_weights}
\alias{update_weights}
\title{Update weights}
\usage{
update_weights(x, weights = NULL, r = 1)
}
\arguments{
\item{x}{A strictly positive numeric vector.}

\item{weights}{A strictly positive numeric vector of weights, the same length
as \code{x}. The default is to equally weight each element of \code{x}.}

\item{r}{A finite number giving the order of the generalized mean.}
}
\value{
A numeric vector the same length as \code{x}.
}
\description{
Factor weights to turn the generalized mean of a product into the product
of generalized means. Useful for price-updating the weights in a
generalized-mean index.
}
\details{
This function generalizes the result in section C.5 of Chapter 9 of the PPI
Manual for chaining the Young index, and gives a way to chain
generalized-mean price indexes over time.

\preformatted{gmean(x * y, w) == gmean(x, w) * gmean(y, update_weights(x, w))}

Factoring weights return a value that is the same length as \code{x},
so any missing values in \code{x} or the weights will return \code{NA}.
Unless all values are \code{NA}, however, the result will still satisfy
the above identity when \code{na.rm = TRUE}.
}
\examples{
x <- 1:3
y <- 4:6
w <- 3:1

# Factor the arithmetic mean by chaining the calculation.

gmean(x * y, w)
gmean(x, w) * gmean(y, update_weights(x, w))

# In cases where x and y have the same order, Chebyshev's
# inequality implies that the chained calculation is too small.

gmean(x * y, w) > gmean(x, w) * gmean(y, w)
}
\references{
ILO, IMF, OECD, UNECE, and World Bank. (2004).
\emph{Producer Price Index Manual: Theory and Practice}.
International Monetary Fund.
}
\seealso{
Other math functions:
\code{\link[=gmean]{gmean()}}
}
\concept{math functions}
